requirements.txt
python-dotenv~=0.19.0
fastapi~=0.79.1
uvicorn~=0.18.3
nonebot-adapter-onebot==2.1.3
nonebot2~=2.0.0b5
websockets~=10.3